+++ affinda/models/_models_py3.py
@@ -9725,6 +9725,10 @@ class JobDescriptionSearchResult(msrest.serialization.Model):
     :vartype search_expression: ~affinda.models.SearchExpressionSearchScoreComponent
     :ivar organization_name: Required.
     :vartype organization_name: str
+    :ivar custom_data: Dictionary of
+     <components·nqbw24·schemas·customdatasearchscorecomponent·additionalproperties>.
+    :vartype custom_data: dict[str,
+     ~affinda.models.ComponentsNqbw24SchemasCustomdatasearchscorecomponentAdditionalproperties]
     """
 
     _validation = {
@@ -9765,6 +9769,10 @@ class JobDescriptionSearchResult(msrest.serialization.Model):
             "type": "SearchExpressionSearchScoreComponent",
         },
         "organization_name": {"key": "organizationName", "type": "str"},
+        "custom_data": {
+            "key": "customData",
+            "type": "{ComponentsNqbw24SchemasCustomdatasearchscorecomponentAdditionalproperties}",
+        },
     }
 
     def __init__(
@@ -9783,6 +9791,12 @@ class JobDescriptionSearchResult(msrest.serialization.Model):
         search_expression: "_models.SearchExpressionSearchScoreComponent",
         organization_name: str,
         occupation_group: Optional["_models.OccupationGroupSearchScoreComponent"] = None,
+        custom_data: Optional[
+            Dict[
+                str,
+                "_models.ComponentsNqbw24SchemasCustomdatasearchscorecomponentAdditionalproperties",
+            ]
+        ] = None,
         **kwargs,
     ):
         """
@@ -9812,6 +9826,10 @@ class JobDescriptionSearchResult(msrest.serialization.Model):
         :paramtype search_expression: ~affinda.models.SearchExpressionSearchScoreComponent
         :keyword organization_name: Required.
         :paramtype organization_name: str
+        :keyword custom_data: Dictionary of
+         <components·nqbw24·schemas·customdatasearchscorecomponent·additionalproperties>.
+        :paramtype custom_data: dict[str,
+         ~affinda.models.ComponentsNqbw24SchemasCustomdatasearchscorecomponentAdditionalproperties]
         """
         super(JobDescriptionSearchResult, self).__init__(**kwargs)
         self.identifier = identifier
@@ -9827,6 +9845,7 @@ class JobDescriptionSearchResult(msrest.serialization.Model):
         self.occupation_group = occupation_group
         self.search_expression = search_expression
         self.organization_name = organization_name
+        self.custom_data = custom_data
 
 
 class JobTitleParsed(msrest.serialization.Model):